NEW: Start port of rectangle geometry
